Move node parse logic into Parsable modules
